(#101) Avoid flicker of login/logout/denied buttons
Currently, the 3 login, logout and denied buttons are shown before that 2 of them are hidden, which causes an unpleasant flicker. This happens because the AngularJS code takes time to kick-in.
Adding the "ng-hide" class to them hide them by default. This way, only the correct button will appear, after a short moment.